aiida-aspherix/pyproject.toml
[build-system]
# build the package with [flit](https://flit.readthedocs.io)
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

[project]
# See https://www.python.org/dev/peps/pep-0621/
name = "aiida-aspherix"
dynamic = ["version"]  # read from aiida_aspherix/__init__.py
description = "AiiDA plugin that wraps the `aspherix` executable "
authors = [{name = "The AiiDA Team"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Development Status :: 3 - Alpha",
    "Framework :: AiiDA"
]
keywords = ["aiida", "plugin"]
requires-python = ">=3.8"
dependencies = [
    "aiida-core>=2.0,<3",
    "voluptuous"
]



[project.optional-dependencies]
testing = [
    "pgtest~=1.3.1",
    "wheel~=0.31",
    "coverage[toml]",
    "pytest~=6.0",
    "pytest-cov"
]
pre-commit = [
    "pre-commit~=2.2",
    "pylint>=2.5.0,<2.9"
]
docs = [
    "sphinx",
    "sphinxcontrib-contentui",
    "sphinxcontrib-details-directive",
    "furo",
    "markupsafe<2.1"
]

[project.entry-points."aiida.data"]
"aspherix.particles" = "aiida_aspherix.data:AspherixPositionsAndType"
"aspherix.parameters" = "aiida_aspherix.data.asx_parameters:AspherixParameters"

[project.entry-points."aiida.calculations"]
"aspherix.insertion" = "aiida_aspherix.calculations.aspherix.insertion:AspherixInsertion"
"aspherix.insertion_compression" = "aiida_aspherix.calculations.aspherix.insertion_compression:AspherixInsertionCompression"

[project.entry-points."aiida.parsers"]
"aspherix" = "aiida_aspherix.parsers:AspherixParser"

[project.entry-points."aiida.cmdline.data"]
"aspherix" = "aiida_aspherix.cli:data_cli"

[tool.flit.module]
name = "aiida_aspherix"
